account for executables on Windows
